refactor: 모든 피스 테스트 Space에 의존적이지 않도록 수정
