+++ src/test/java/chess/piece/KingTest.java
@@ -1,11 +1,8 @@
 package chess.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.domain.chessBoard.Space;
 import chess.domain.piece.Color;
-import chess.domain.piece.EmptyPiece;
 import chess.domain.piece.King;
 import chess.domain.piece.Piece;
 import chess.domain.position.File;
@@ -20,50 +17,29 @@ public class KingTest {
     @DisplayName("상하좌우로 한 칸 움직일 수 있다.")
     void should_move_up_down_left_right() {
         Piece piece = new King(Color.WHITE);
-        Space space1 = new Space(piece, new Position(File.a, Rank.ONE));
-        Space space2 = new Space(new EmptyPiece(), new Position(File.a, Rank.TWO));
+        Position from = new Position(File.a, Rank.ONE);
+        Position to = new Position(File.a, Rank.TWO);
 
-        space1.movePiece(space2);
-
-        assertThat(space2.getPiece()).isEqualTo(piece);
+        assertThat(piece.isMovable(from, to)).isTrue();
     }
 
     @Test
     @DisplayName("대각선 방향으로 한 칸 움직일 수 있다")
     void should_move_diagonal() {
         Piece piece = new King(Color.WHITE);
+        Position from = new Position(File.a, Rank.ONE);
+        Position to = new Position(File.b, Rank.TWO);
 
-        Space space1 = new Space(piece, new Position(File.a, Rank.ONE));
-        Space space2 = new Space(new EmptyPiece(), new Position(File.b, Rank.TWO));
-
-        space1.movePiece(space2);
-
-        assertThat(space2.getPiece()).isEqualTo(piece);
+        assertThat(piece.isMovable(from, to)).isTrue();
     }
 
     @Test
     @DisplayName("두 칸 이상 움직일 수 없다")
     void should_not_move_more_than_two_space() {
         Piece piece = new King(Color.WHITE);
+        Position from = new Position(File.a, Rank.ONE);
+        Position to = new Position(File.a, Rank.THREE);
 
-        Space space1 = new Space(piece, new Position(File.a, Rank.ONE));
-        Space space2 = new Space(new EmptyPiece(), new Position(File.b, Rank.THREE));
-
-        assertThatThrownBy(() -> space1.movePiece(space2))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이동 규칙을 위반한 움직임입니다.");
-    }
-
-    @Test
-    @DisplayName("이동할 위치에 상대 말이 있고 잡을 수 있으면 이동할 수 있다")
-    void should_move_when_target_space_has_other_color_piece_and_catchable() {
-        Piece piece1 = new King(Color.WHITE);
-        Piece piece2 = new King(Color.BLACK);
-        Space space1 = new Space(piece1, new Position(File.a, Rank.ONE));
-        Space space2 = new Space(piece2, new Position(File.b, Rank.TWO));
-
-        space1.movePiece(space2);
-
-        assertThat(space2.getPiece()).isEqualTo(piece1);
+        assertThat(piece.isMovable(from, to)).isFalse();
     }
 }
