+++ src/test/java/chess/piece/RookTest.java
@@ -2,9 +2,7 @@ package chess.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.chessBoard.Space;
 import chess.domain.piece.Color;
-import chess.domain.piece.EmptyPiece;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Rook;
 import chess.domain.position.File;
@@ -19,24 +17,9 @@ class RookTest {
     @DisplayName("상하좌우로 칸 수 제한 없이 움직일 수 있다.")
     void should_move_up_down_left_right_unlimited() {
         Piece piece = new Rook(Color.WHITE);
-        Space space1 = new Space(piece, new Position(File.a, Rank.ONE));
-        Space space2 = new Space(new EmptyPiece(), new Position(File.a, Rank.THREE));
+        Position from = new Position(File.a, Rank.ONE);
+        Position to = new Position(File.a, Rank.THREE);
 
-        space1.movePiece(space2);
-
-        assertThat(space2.getPiece()).isEqualTo(piece);
-    }
-
-    @Test
-    @DisplayName("이동할 위치에 상대 말이 있고 잡을 수 있으면 이동할 수 있다")
-    void should_move_when_target_space_has_other_color_piece_and_catchable() {
-        Piece piece1 = new Rook(Color.WHITE);
-        Piece piece2 = new Rook(Color.BLACK);
-        Space space1 = new Space(piece1, new Position(File.a, Rank.ONE));
-        Space space2 = new Space(piece2, new Position(File.a, Rank.THREE));
-
-        space1.movePiece(space2);
-
-        assertThat(space2.getPiece()).isEqualTo(piece1);
+        assertThat(piece.isMovable(from, to)).isTrue();
     }
 }
